calibration works, added info display

+++ Core/Src/Touch/Touch.h
@@ -3,6 +3,7 @@
 #include "main.h"
 #include "stm32f4xx_hal.h"
 #include "ILI9341_Driver.h"
+#include "stdio.h"
 
 #define TP_PRES_DOWN 0x80
 #define TP_CATH_PRES 0x40
@@ -62,4 +63,4 @@ uint8_t TP_Get_Adjdata(void);
 void TP_Adjust(void);
 uint8_t TP_Init(void);
 
-void TP_Adj_Info_Show(uint16_t x0,uint16_t y0,uint16_t x1,uint16_t y1,uint16_t x2,uint16_t y2,uint16_t x3,uint16_t y3,uint16_t fac);
+void TP_Adj_Info_Show(uint16_t x0,uint16_t y0,uint16_t x1,uint16_t y1,uint16_t x2,uint16_t y2,uint16_t x3,uint16_t y3, double fac);
